+++ annotations/src/main/java/arez/annotations/ContextRef.java
@@ -3,7 +3,6 @@ package arez.annotations;
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Target;
-import org.realityforge.anodoc.Unsupported;
 
 /**
  * Marks a template method that returns the {@link arez.ArezContext} instance for component.
@@ -21,7 +20,6 @@ import org.realityforge.anodoc.Unsupported;
  */
 @Documented
 @Target( ElementType.METHOD )
-@Unsupported( "This is largely experimental and intended for framework users rather than casual users of the framework" )
 public @interface ContextRef
 {
 }