+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Unreleased
 
+* **\[annotations\]** Remove the `@Unsupported` annotation from all the annotations within the
+  `arez.annotations` package as all of the annotations are sufficient stable to be supported
+  going forward.
+
 ### [v0.91](https://github.com/arez/arez/tree/v0.91) (2018-06-13)
 [Full Changelog](https://github.com/arez/arez/compare/v0.90...v0.91)